Extract file indexability partial
